Debugging - COB working
Chased down a couple of bugs: first, I specified "Tribal or Equivalent" for a field value in one place but "Tribal Area or Equivalent" further down; my tribals were not found. Then, I was assuming I could check for adds based only on Dec string when I know that is not unique. Both fixed. Working. Don't breathe on it.